requirements.txt
mutagen
click
pyyaml
